iot-cloud-project/sensorApi/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
google-cloud-bigquery>=2.0.0
gunicorn>=20.0.0
